Fixup routing and handler stubs
The routes are now generated in the API files, which is the only way
to guarantee that they are in fact generated.

The stub handlers are now generated out of the way in a supporting file
called stubs.go. This isn't perfect, but it works OK. I haven't relied
on the .swagger-codegen-ignore mechanism because we know that the stubs
will always be replaced by code that we must not overwrite.