Remove the Gordian service definitions
Now, an APIClient simply contains a Configuration (not a pointer).

The various API service types are simply type aliases for the
APIClient type because there isn't any need for anything extra at the
moment.

The *Api fields in the APIClient that were pointers to the various API
service types have been replaced by *ApiService methods of the
APIClient that return that same pointer.