Implement a package doc comment.
The partial_header was being picked up as a package doc comment
because of its location. I moved a doc comment instance of it into its
own file and moved the partial_header to just below the ``package``
statement so that it won't be picked up as a doc comment.

Reformatted/rewrote the comment in the partial_header to meet the
style guidelines for a Go doc comment.