Initial support for separate model and api packages
This is initial support for my particular use case where I have
implemented the models by hand, but want the API to be automatically
generated for me. I use --model-package and --api-package to specify
the two package names and I use -DsupportingFiles and -Dapis to skip
generation of the models.

This made the partial_header untenable because it relied on there only
being one package name, allowing the header to be shared across the
models and the API. Since Go expects that sort of comment to only go
into one file as a package comment, I did that.

I also implemented the Go comment marker for automatically generated
source code since I was messing with all that commentary anyway.

Fixes #22
Fixes #23